Added support for multiple content entries per module.
The problem:
Users may customize the location of the Gradle build folder. If
such build folder is outside the root folder of the Android project
('project' in Gradle terminology,) Studio will fail to import the
project as a module, because it does not allow source folders
outside of the module root folder ('module' in IDEA terminology.)
The build folder contains generated folders.

An additional problem is that today I learned that a module may
have multiple content roots (a "content root" is just a folder that
is the parent, in the file system, of the source, test and
resource folders of a module.) Studio only supports one content root
per module.

Solution:
This CL allows multiple content roots per module. For Android projects
this change does not make a difference, because we only have one
root directory. This CL will be useful once we use the new Gradle
model that exposes the build directory (CL
https://android-review.googlesource.com/#/c/83053/ ). Once we have
that in place, we can have a separate content root for the build
directory it is outside of the project's root directory. This will
fix a severe and quite popular bug in Studio.

Another benefit of this CL is that the Java model that we get from
Gradle already supports multiple content roots, but we were discarding
all of them except for the first non-null one.

Change-Id: I3d2d4f903a01bb2def947e59c956b777949d4cfb