Update build.gradle when module order entries change
Adds a listener for module order entries and updates build.gradle
files when it sees new entries appear there. This will fix bugs
where some IDE operations (such as the quick fix for adding
a module dependency when importing a class that needs it) update
the IntelliJ side of the project structure but not Gradle.

This CL includes a facility for querying the build.gradle file
to see if it has a dependency that "matches" a given dependency;
it will be useful for merging together build file dependencies
in other contexts.

Bug: 56178
Change-Id: I9063ff70aed17c377aa0703dade061305d344cbc