Added support for source folders outside of module's root directory.
This CL collects all the source folders that are not inside a module's
content root (the module's root directory.) As it collects these
orphan source folders, it creates a list of common folder ancestors.
These list of common ancestors is later on used to create additional
content roots for the orphan source folders.

Fixes:
https://code.google.com/p/android/issues/detail?id=56367

Change-Id: Ic24c048eeb4a8b2c3cc135acbb66814e464e6f58