Fixed bug where "Compile Module" and "Make Module" actions won't do anything.
The problem:
Selecting the project's root directory and then selection either "Compile
Module" or "Make Module" actions won't do anything. The cause of the problem
is that when this directory is selected and if this directory does not
map to a Gradle project, Studio should build all the modules in the project.

To know if Studio should build all the modules when the project root directory,
the module at the project's root level should not have the Gradle facet and should
have the same path as the project. We were incorrectly checking for a match
between the name of the module and the name of the project.

Change-Id: I4c67a0c4fa5909f2777d4878204f818d566d422a