Save setting for "Use same device for future launches" checkbox
The device chooser dialog has a checkbox that allows users to
specify whether the same device should be used for future launches.
This setting was being saved when the checkbox was enabled, but not
when the checkbox is disabled (after it was enabled at some point
in the past, and the previous device in use isn't there anymore).

This CL fixes things so that the deselection is persisted.
Fixes https://code.google.com/p/android/issues/detail?id=66155

Change-Id: If65540795ba6e76332bf512900173a6709acbb11