screenshot viewer: save updates to backing file
The screenshot viewer uses the IJ image viewer to display images.
One of the actions provided by this viewer is to open the file in
an external viewer. This relies on the image currently displayed
also being present in the backing file.

This CL makes sure that any time user does a processing action
such as reloading an image or applying device frames, the actions
update not just the in memory copy but also save it out to the
backing file, so that viewing them in an external file continues
to work.

Fixes https://code.google.com/p/android/issues/detail?id=65729

Change-Id: If366a788619e205a2b8f810aac100ea5ef41195d