Fixed bug where quick-fix did not update Android plug-in version in build files.
Problem:
Android Studio shows a balloon notification when opening a project using an old
version of the Android Gradle plug-in. This balloon has a the link
"Fix plug-in version and re-import project". When clicking the link nothing
happened, and an error was displayed.

Solution:
Modify the underlying build file inside a WriteCommandAction.

Change-Id: I25b6c8885a56177a46b6664ee59cb454b2b1940b