Now we add build dir as a separate content root if it is outside the module's root dir.
CL https://android-review.googlesource.com/83486 added the ability to have multiple
content roots for a module. This comes handy when users defined the build dir outside
of the module's root dir (we did not allow this before.) The problem is that if the
build dir is inside the module's root dir we are creating a new content root anyway,
which changes the way the "Project" view looks, without adding any value.

This CL only adds a separate content root for the build dir if the build dir is
outside of the module's content root.

Change-Id: Ifd6617577aabb95f145512b8170c76fff0c8031b