Move clear logcat action to the top of list of actions
This CL moves the 'clear logcat' and the 'scroll to end' actions
to the top of the list of actions for the logcat toolwindow. This
avoids the clear action getting folded into the overflow area
when the window is not sufficiently tall to fit all the actions.

Fixes https://code.google.com/p/android/issues/detail?id=66626

Change-Id: I8d261bab66e74955392904af46d98b535d09d015