Generalize static analysis to support both the Nurik Master-Detail and SimpleMail idioms
This CL adds a more principled/generalized 'evaluator' to perform the static
analysis of Activities and Fragments. In this generalization we are effectively
propagating the known values of two method calls, namely:

* findViewById()
* findFragmentByTag()

up through their enclosing AST's so as to be able to figure out which branches
of conditionals, like:

if (findViewById("foo") == null) { <launch activity>} else { <something else>}

will be executed when the app is run on different devices.

The analysis is complicated by assignments to fields and local variables and this CL
greatly increases the number of cases that are handled correctly - providing a sufficiently
general method to handle both the Nurik Master-Detail and SimpleMail idioms correctly.

Change-Id: I3c397b30657e7ca396cff382769d5c9fe10e93db