Fix for NavigationModels not correctly preserving Activity/Menu locations between saves
Location information can't be stored with the Activity and Menu states themselves.
This is how things were in the previous version; though it's now clear that using States
as keys in HashMaps, while simultaneously expecting to mutate their locations is a
modelling error - which was confusing the editor's internal caches.

This CL makes States immutable so they can be used as keys, as before.
Locations are now stored in a separate map owned by the NavigationModel.

Change-Id: Ib93857eb05f9beaba077b6da532e050797687c79