Split CompilerOutputPathModuleCustomizer into "android" and "java" implementations.
This CL:
- Moves the hard-coded output paths for Java lib modules from
  CompilerOutputPathModuleCustomizer to JavaModel (where they
  belong)
- Adds a new CompilerOutputModuleCustomizer that sets a module's
  output paths from a JavaModule

We still have the issue that we do not get output paths for
Java lib modules from Gradle. We are working on that, and changes
need to be made in IDEA. Once we have those changes in place,
I'll remove the hard-coded paths from JavaModel and replace them
with the ones coming from Gradle.

This CL cleans up the code in preparation to the upcoming changes
in IDEA to deal with compiler outputs for Java lib modules.

Change-Id: I548e150e14d2283048a405995d3ba94210459f21