Fixed bug where custom project structure was not supported.
Problem:
When users configure the project structure in build.gradle and
specify that a folder (e.g. "src") is both a Java source
folder and a resource folder, Studio will mark it as a
"resource" folder only, and .java files were not recognized as
Java code. Here is an example of such customization:

    sourceSets {
        main {
            manifest.srcFile 'AndroidManifest.xml'
            java.srcDirs = ['src']
            jniLibs.srcDirs = ['libs']
            resources.srcDirs = ['src']
            aidl.srcDirs = ['src']
            renderscript.srcDirs = ['src']
            res.srcDirs = ['res']
            assets.srcDirs = ['assets']
        }
    }

Solution:
The root of the problem is that Studio first marks folders as
"source" and then as "resource", and in this case, the "source"
mark gets lost. The fix is to do the opposite: mark "resource"
first and then "source" (source has higher priority than
resource.)

Fixes
https://code.google.com/p/android/issues/detail?id=66758

Change-Id: If58017bd10629fb68786712547d469c8573161f8