Enable run configurations to listen to refactorings
Run configurations can implement the RefactoringListenerProvider to
listen to refactoring changes in the IDE and update their state if
necessary.

The Android run configuration listened to this event and would update
the class name saved in the run configuration if it happened to change.
However, it did not update itself if the containing package changed.
This CL fixes that. In addition, it also makes the Android test
run configuration implement the same interface and listen to changes
in the package, class or method names and update as necessary.

Fixes https://code.google.com/p/android/issues/detail?id=64916

Change-Id: I46daa6993056ab4134a778f49fb8fd9f441d949d