Defensive checks for ManifestInfo
ManifestInfo can only be obtained for modules with the AndroidFacet.
This condition was enforced in MergedManifestInfo class but not in
the PrimaryManifestInfo. This CL moves the check into the factory
method.

In addition we make the public methods of ManifestFile class to
be synchronized in order to protect against possible corruption
due to multi threaded access.

Change-Id: I332d1559de05ec7f387cd3b7c1bd83fa36e771e1