Fully qualify activity alias names
Earlier we used to obtain the activity alias name from the merged
manifest where it was fully qualified. Now that we are reading the
manifests directly, it is possible that they aren't fully qualified
and this leads to trouble if the package name is customized by
Gradle.

This CL attempts to read the package name as defined in the manifest
that contains the activity alias and use that to fully qualify
the activity alias.

Fixes https://code.google.com/p/android/issues/detail?id=67010

Change-Id: If2f5e77da6280742bc033a7f88c9905decb39c53