Fixed wrong assumption that a Gradle sub-project must have a build.gradle file.
Studio used to create IDEA modules for Gradle sub-projects only if a sub-project
has a build.gradle file. It turns out that having Gradle sub-projects without
build.gradle files is valid (e.g. configuration can be done at the top-level
parent project.)

The problem with this assumption is that if sub-project B depends on sub-project
A, and sub-project A does not have a build.gradle file, B is going to link to
the compiled jar file of A, instead of the IDEA module. Studio flags this as an
import errors, and users may not be able to navigate from classes in B that use
classes in A.

Change-Id: I2617fc929d18f23447d658ed2f4ae084b64847af