Fixed bug where physical Gradle sub-project path was derived from logical Gradle path.
Problem:
IDEA by default, when importing a Gradle project, deduces the physical path (in the
file system) of the IDEA module to create from the Gradle sub-project's logical path.
For example, if the logical path is dir1:dir2:app, the deduced physical path would
be ${projectDirPath}/dir1/dir2/app. This mechanism is not reliable, because in
settings.gradle it is possible to map logical sub-project paths to any physical path,
for example:

include ':app'
project(':app').projectDir = new File(rootDir, 'test/app')

This mapping breaks IDEA's assumption, because :app is mapped to test/app.

Solution:
This CL fixes this problem by getting the physical path of a Gradle sub-project from
a GradleBuild. GradleBuild is an additional model that we query from Gradle and has
a lot more information about sub-projects, including the actual physical path.

Change-Id: Ifa59c4810615fa27f5f8d5753bc89ea85374ff39