Move .navigation directory to top level (next to .idea directory)
This CL allows the .navigation directory to be moved to the top of the project
structure. The Navigation Editor still really applies only to a single module and this
CL includes code to find that module from the project - by assuming that
the project has only one module with an AndroidFacet. If the number of
modules that have an AndroidFacet is not equal to one, the Navigation Editor
aborts.

Change-Id: Ifff14802e24d5d79316acf820f0c8f4208e5c08e