Improved IDE behavior when project sync errors found.
Problem:
When there are project-sync-related errors, and these
errors are shown in the "Messages" window, the editor
notification saying "Gradle project sync failed" did
not show up. Also, the "gen-source" build was being
triggered, which is unnecessary, since we have project
setup problems.

Solution:
When this type of errors are detected, the IDE:
- Shows the "Gradle project sync failed" editor
  notification
- Does not do any post-sync processing, including
  any "gen-source" builds.

Change-Id: I6727d8436f7aabb6fe0d5f3f36320afc43817d8c