Now we warn if SDK does not have platforms installed when setting the default Android SDK.
Scott reported this issue and had a reproducible use case (which made everything
a lot easier!)

Studio's behavior was confusing because when no platforms were found for a given
Android SDK path, it failed silently, and it seemed that it never stored the given
path.

In this CL we at least warn the user about the SDK not having platforms. Ideally
we would include a link saying "Open SDK Manager" but at this point is not feasible.
We need deeper changes to make this happen. With this CL we at least warn the user
(as supposed to fail silently) and I will find a way to offer opening the SDK
Manager when this error occurs.

Change-Id: Ifa2debd1cc1288795d7de127743bd18cd2aaba77