Validation on project import.
When importing a project, open it immediately
so we can see the messages window and then run Lint on it
before project import begins. This gives us an opportunity to
detect errors and provide helpful error messages if the import
fails. It also gives us a framework to write future validation
checks for the build files that will work not only for import
but for the build file editors as well.

Change-Id: Ife4a5f226d6091199f20067957e9a0427246ba14