Add resource type annotation, and rewrite nullness/typedef annotations
This changeset rewrites the bundled annotations for the platform
SDK to match the ones from the support library, to unify handling for
the annotations from the SDK with annotations found not just in the
support library but in any AAR supplying its own annotations. It also
removes all the inferred @Nullable annotations; these were often too
aggressive. It preserves the @Nullable annotations that have been
manually added to the framework source code so far.

Change-Id: I4836bb0536914855b98d4b8b3b129b3868c02c05