New Project Structure infrastructure.
Problem:
The current "Project Structure" in Android Studio reuses
the same infrastruture as IDEA. This infrastructure has, under
the covers, many editors actively listening for changes
in the the project structure that conflict with the changes
we made in our project structure (e.g. there are many bugs
regarding objects "already disposed" errors.
The default infrastructure also has a daemon that constantly
checks for errors like "libraries not being used" and displays
such errors in the dialog. The problem with this is we cannot
override these errors, and the "quick fix" for these errors
are not Gradle-aware: instead of modifying Gradle build files,
they modify .iml files.
It is also difficult to customize the look of the "Project
Structure" dialog, to remove components that we don't need
or change layouts of things.

Solution:
This CL introduces a new "Project Structure" infrastructure
written from scratch that does not reuse IDEA's. All the problems
stated aboved are fixed by this change.

This CL:
- Adds the new "Project Structure" configurable (the contents of
  the dialog.)
- Changes the layout of the contents. Instead of 3 panes, we use
  2, consolidating the "SDKs" entry with the "Modules" entry. We
  save space needed for the editors.
- The "Project Structure" only performs sync when really needed.
- The dialog has a notification when sync is in progress.
- The dialog remembers UI state, including previously selected
  configurables.
- Uses JVM arg "use.idea.projectStructure" to fall back to IDEA's
  project structure (for debugging.)

TODO:
- Add "remove module" action.
- Use icons for different module types (app, lib and java lib.)
- Replace table-based editors with master/detail ones.

Fixes
https://code.google.com/p/android/issues/detail?id=65595
https://code.google.com/p/android/issues/detail?id=68095

Change-Id: I451debdd90bf5bcf466004da5fef006ff3e86c7b