Resolver for the android configuration closure
This CL improves on the existing resolver for the android
configuration closure in gradle build scripts.

The earlier version relied on knowing the class that contributed
each of the configuration extensions inside the android closure.
As a result, it only worked for the classes it knew about.

With this CL, the resolver only knows what class the android
extension maps to. All extensions within the android block
are resolved by looking for matching methods within the class
corresponding to the android extension. It also assumes that
the Action class from org.gradle.api indicates that something
is a closure, and a NamedDomainObject indicates that a new
domain object of given type will be created.

Change-Id: I1d440ace3e5d6018eb30e526a283fe856c13fafb