Now we pass option --configure-on-demand to Gradle builds.
This option tells Gradle to configure only the sub-projects
(or modules, in IDEA lingo) that are going to be built, instead
of all the sub-projects in a given project. The result is
faster builds for multi-project projects.

Gradle docs:
http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:configuration_on_demand

This CL adds a new option in the Compiler > Gradle preference
page to allow users to enable/disable this option (it is
"enabled" by default.)

Change-Id: Ia6ccb72a653b0c96e8fca191d9058023b10a62cf