Fixed bug where IDE did not create SDKs for platforms installed via SDK Manager.
This is how I found the bug:
1. Have a SDK with platform android-19 only
2. Have a project that uses android-18
3. Studio complains that the platform is not installed and provides
   link to open SDK Manager
4. Install missing platform
5. Restart Studio
6. Studio keeps complaining that android-18 is missing, even though
   it was installed

The problem is that we don't create any missing SDKs on start up. Ideally
we'll create SDKs in the IDE without a restart, but the SDK Manager runs
in a separate process and we cannot attach a listener to know when new
platforms are installed (we are working on an in-process SDK Manager, BTW.)

This CL creates IDE SDKs at startup, fixing this issue.

Change-Id: I88dab140a3af21a618c275459c8a04381ec0df9e