Fix bug where layout editor displays "Rendering..." progress forever
The layout editor was using a shared, per project, merging queue to
queue up and merge pending render requests.

On render requests, the layout editor would start a new 500ms alarm,
which posts a "Rendering..." progress bar in the editor when it
expires.  It would then queue a render request into the merging
queue. And in the merging queue's processing method, it would call
into layoutlib to render, and when all that was done, it would cancel
the alarm and hide the progress UI (if the alarm had expired). This
all allows the layout render to show progress if it takes more than
500ms but to just show the results immediately if it's shorter.

However, there have been reports of occasional cases where the
"Rendering..." message stays up forever. It turns out that this is
because the session queue is shared between files. Consider what
happens if we switch files right after a rendering request has been
started. We first schedule an alarm for file A, and add a render
request (which will cancel the alarm when done). But then we switch
files and file B also files a render request, which is now merged with
the existing pending request, and one of the files will *not* run its
code to remove the alarm.

Using a shared merging queue doesn't work well for the layout editors
the way they're currently implemented: it's okay to collapse multiple
rapid requests for the same file, but not for two separate files.

Change-Id: I62461d8ac96475bea1b60953973bd1dc3ac95eb9