Make layout preview respond immediately to configuration changes
The layout preview window uses a timer to delay updates, such it only
updates itself after you stop typing in the editor and wait for a
while (800ms or so). Similarly, when switching files, it only updates
itself after a while (such that the source editor is immediately
responsive). However, when you're configuring layout by clicking on
actions in the configuration menu (such as switching devices or
toggling between portrait and landscape), this delay just makes the
rendering seem sluggish. This CL addresses this by making these
operations apply right away.

Second, it tries to preserve the XML versus Layout Editor mode for
automatic file switches. Let's say you have both layout/foo.xml and
layout-land/foo.xml. If you're editing layout/foo.xml with the layout
editor, and you click on the orientation toggle to switch to
landscape, it will switch to layout-land/foo.xml since that's the
correct match, and with this CL it will also switch to the layout
editor for that new file. Similarly, if you were editing the file as
an XML file, it will switch the target file to the same mode.

Finally, it moves the layout editor's editor ID constant over from
RenderPanel to AndroidDesignerEditorProvider where it belongs, now
that the android-designer sources can be accessed directly from the
editor code.

Change-Id: If2cd17565dca57680aa12de405c31827341cf67b