Pass the menu from tools attribute in layout file
The changeset passes makes the Action Bar menus passed to layoutlib
configurable by using the tools:menu attribute in the top level layout
file.

This requires a reference to the RenderService in the ProjectCallback.
Since the callback is used per file and is not project wide, it is more
sensible to call it LayoutlibCallback. The interface however is not
renamed since that will break compatibility with older versions of
layoutlib.

Change-Id: Ie57ee406b7977dca0576177dc3a62dbcf0484fda