Fix bug where "..." button to select SDK location was disabled.
This is happening in the "Project Structure." After the last IDEA
update, it seems like the "..." button becomes disabled if the
text field of the browse component is not editable. Such text field
should not have been editable from the first place. This CL fixes
that and also, by making the text field editable, the "..." works
again.

Change-Id: I61a14d38c7299f4d25cedb3d39bfb981a0612f8a