Fix default layout orientation picking code
First, restore the behavior that when you switch between devices in
their default orientation, we switch to the default orientation for
the target device as well, as long as there is no layout file
switch. In other words, if you view a layout with a Nexus 4 in
Portrait orientation (its default), and you switch to a Nexus 10, we
also switch to Landscape orientation (its default), unless there is a
layout-land/ file that is a better match.

Second, fix some bugs in the code to handle state/device matching when
viewing files in a layout-land/ folder.

Change-Id: I3309b11cd383a998356b17b4b540b6a0d3f66b0a