Initial implementation of the "Project Profiles" dialog.
This CL introduces the "Project Profiles" dialog. It is not
hooked up into project sync yet. This dialog allows users
to select a subset of modules to import into Studio based on
selection variant conflicts found during project sync.

The dialog shows all modules existing in a project, all the
selection variant conflicts and the details of each conflict.
Using this information, users can select the modules to
import, and while doing that, the dialog will indicate whether
a conflict has been resolved based on the current selection.

This CL does not change any behavior in the IDE, because the
dialog is not hooked up yet.

Change-Id: I2a854af9f74ec07311775ea4d63bfa0e1cc6e952