Add ability to create from a github-hosted template
This change adds a NewFromGithubWizard which can download
a github repository and scan it for templates and gradle-based
projects. If templates are found, they can be inflated with the
standard template mechanism. If the repository instead contains
samples, then the wizard will offer to import them as modules
into your project.

The UI action is disabled by default but may be enabled by adding
setting the system property "enable.import.from.github"

Change-Id: I73ea18f833095e58f40337e050aecbf4fb991c8d