Fixed bug where editor notifications cannot be refreshed if project structure changed.
This bug does not happen often. I was able to reproduce with a
project that was successfully imported, made some changes that made
the project structure change (in my case, removes all the modules)
and then an exception was thrown when updating editor notifications
because the modules that existed before the project sync are now
disposed.

Change-Id: I48089eb8cbb00a49bb41cf86aae16408b77f1434