Update the "SDK valid" error message while user edits textfield
After upgrading my JDK I had an invalid JDK path. I copy pasted
the new path into the text field but it still displayed an error
message stating that the JDK path was not valid. This is because
validation only occurred after using the file chooser / browse
button. This CL adds a listener to the text field and updates
the error message after each edit.

Change-Id: Ica3ed3d6c94ae0d9e9df1f3096af38a143fa4044