Allow new project creating in existing directory
Previously we did not allow creating of a new studio project in an existing
directory. We now allow creation of the new project in an existing directory
if it has no children, or, if it does have children, we warn the user but
do not prevent them from going forward.

This allows project creation in cloned git repositories.

Fixes https://code.google.com/p/android/issues/detail?id=56159 and
https://code.google.com/p/android/issues/detail?id=67042

Change-Id: If15e34d6c45ae6d0c651fda21984c50918ed4d4b