Fixed bug where adding a stats record freeze the UI.
The method StudioBuildStatsPersistentComponent#addBuildRecord was
being executed in the UI thread, making the UI freeze when having
a slow Internet connection (this method checks some URL.)

This CL executes this method in a separate thread. I'm not sure
this is the best way to do this, but at least it works well when
having a slow Internet connection.

Change-Id: I077bd4d281cd35aed87a938b35862720942f653d