device panel: don't trigger selection listeners on certain updates
The device pane in the android toolwindow shows a list of clients
(running processes). When the list of processes changes, we remove
all the elements from the current model, and add the new set of
processes to the model, while still maintaining selection if the
currently selected client is still present in the new list.

However, the selection listeners were triggered during this process
since the model changed. This CL adds a flag to distinguish between
real user selection events and the above case where it is a internal
model change.

Fixes https://code.google.com/p/android/issues/detail?id=65533

Change-Id: If8fbeb6641bf48c1136d19909c131f81d4e066e4