Fixes a bug in saving NamedObject instances
Fixes a bug I found in working on Project Structure where I saw
an exception while trying to rename an object that had more than
one existing parameter. This was happening because it was
reformatting a Groovy closure too early in the process and a
reference to the closure it was using became stale before it
was done with it. I added a test to avoid future regressions.

Change-Id: Ie64d540ab0b3190a7cf6c9f5e8cd91afa3ca2431