Master/Detail view for Project Structure
In the Project Structure dialog, for named objects (e.g. build types
and flavors but not dependencies), instead of showing items in a
giant 2-dimensional table, show a master/details view where there is
a list of items on the left-hand side with + and - buttons to add
and remove items; clicking on an item in the list shows details in a
2-column table on the right-hand side. This is a less intimidating
view and is easier to navigate for items that have a large number of
properties.

Bug: 68549
Change-Id: I84866d0bc5b81a267f0d29e40c926506951d3be5