Fixes a bug in removing properties from named objects
Fixes a bug where if you tried to remove a property from a named
object (by settings its value to null), it didn't remove that
property from its internal representation, but instead gave it a
null value in its internal key/value map. Through a cascade of
failures this would cause the entire object to be removed from
the Gradle file when writing it out. Added a test to detect this
condition.

Change-Id: I6807e516e6baf1060ee6012dd60fec42018fff00