navigator: package view fixes
This CL fixes a bunch of issues in the package view for
Gradle based Android projects. See
https://code.google.com/p/android/issues/detail?id=68832
for a list of issues.

In particular, in the packages view, the node representing
a module is replaced with a new node, called the AndroidModuleNode.
This node simply overrides the existing package view node for a
module to add some categorization for different source types (the
existing view does not distinguish between java or res, it treats
everything under a source root as a java package).

Manifest files weren't displayed in the package view earlier since
they were outside a source root, but they are displayed now.

Just like the existing package view, the update view also does
not display files that are not in an active source root i.e. files
not enabled for the current variant aren't displayed.

Change-Id: I47d21b5776f40f874b958b99d39c97147e29918a