Track active wizard path in framework, not in each wizard path.
Previously all WizardPath methods were called for all registered paths and
needed to check whether the path was active. This relied on an enum in
NewModuleWizardState, making it more difficult to integrate new paths. Now
the framework keeps track of the active wizard path and calls the update(),
createModule() and isStepVisible() methods only for the active path.

Change-Id: I24d1b364d4bc0b3bac58d2af582c2899298a98fc