Deal with build file reference types
Makes it smarter about reference types in build files. These are
properties such as signing configs where you set up named objects
in one block and then insert Groovy references to them later in
other properties (such as build types or flavors). This CL changes
the handling of them in Project Structure from dumb strings to
a popup that shows you the values it knows about. It also fixes
a bug that would add the signingConfig block in at the end of
the android closure, which would cause unresolved reference
errors -- yes, the order of properties in Gradle build files matters.

Bug: 61687
Bug: 67761
Change-Id: I9d8b74efb141323b55a9ff21a77113d8fae8edec