Speed up included-in search
When the user right clicks in the layout editor, the context menu computes
dynamic actions offering to show the current layout included in another layout,
if applicable. This relies on a usage search of the current layout file.

This CL makes this search faster by specifying a specific scope rather than the
default element scope, which includes both libraries and all file types; we only
need to search the project itself and we only need to look in XML files.

Change-Id: I7d9c929c8a42e5f97c8fc4793b6ccf2535d0adbb