Fix include mask visual artifacts
When viewing a layout rendered inside another included layout, the
including layout is shown behind a semi-translucent mask. However, in
some cases there could be a visual gap. This happened because the mask
is painted as four (or more, in the case of merge tags with multiple
children) separate rectangles which need to abut exactly. Since these
rectangles were mapped to a translated and scaled view port, floating
pointing rounding errors sometimes resulted in these rectangles ending
up an extra pixel apart, which made for a highly visible gap.

This changeset fixes this by performing the rectangle subdivision in
the screen coordinate system rather than the view model coordinate
system, which prevents the divided rectangles from being
scaled/truncated separately.

Change-Id: I755062291123385d8f6abf9f3837827cc45cdda9