Sort goto declaration targets by resource folders, and add tests
When there are multiple goto declaration targets for example in string
translations, the order they were presented in appeared random (tied
to index hash keys perhaps?). This changeset adds explicit sorting
such that the base folder is preferred, and the remaining are sorted
alphabetically.

This also adds unit tests for goto declaration, as well as for Find
Usages: The existing unit tests only asserted the number of usages
found; this tweaks the unit tests to describe the usages by pointing
to the relevant code elements, and asserts that these descriptions are
as expected. Makes it easier to validate that the expected results are
really the ones we need. (Prep for work on the handling of declare
styleable attributes)

Change-Id: Idd6a734d9b92161e2a432a9c523c08102f23a9f5