Uses comboboxes in Project Structure for API levels
In the Project Structure, for fields such as API levels where we
can supply a number of reasonable values based on what we find in
the user's SDK, show those choices in a combo box. The widget allows
the user to enter a custom value if the built-in choices don't
suffice.

Also adds some new options to the general module properties tab.

Bug: 61690
Bug: 61874

Change-Id: Iedd8d0dee1a77ab8f2f174f918b6d9cabebd3dec