Added conflict navigation buttons in "Build Variants" window.
Currently, the "Build Variants" window has a "Show Conflicts"
hyperlink that takes the user to the "Messages" window. From there
it is not easy to navigate or find "variant selection" conflicts.

This CL adds 2 buttons to the "Notifications" panel in the
"Build Variants" window to navigate to "Next" and "Previous"
conflict.

Change-Id: I5d8cc8723d9a793ac35a324cfbaade0823460e17