Project structure panel for project-global properties
Adds a panel in project structure for project-global properties,
generally things found in the root-level Gradle file: plugin version
and repository for all modules, and the library repository for all
modules. Also adds a field to edit the Gradle wrapper version even
though that's in gradle-wrapper.properties and not a Groovy build
file; there's some plumbing to allow the use of BuildFileKey
infrastructure for that.

Bug: 69404
Change-Id: I691b4eb74f5dc3cc411fe9e8907a2f4ba37d9ef1