Speed up platform render resource loading
This changeset speeds up layout rendering initialization, by tweaking the
code which reads in framework resources (where a majority of time in layout
setup is performed). Some simple benchmarks shows this change to speed up
initialization from around 5000ms to around 800ms for the default locale,
and to around 2000ms when rendering platform locales.

This speedup is caused by

(1) Skipping initialization of locale-specific value folders when not
    needed; there are about 140 locale specific folders, and about
    half of those contain a couple thousand strings. We now omit
    locale folders unless you are rendering for a specific locale.

(2) Skipping some other folders we don't need for designtime rendering
    (such as 85 raw- folders and 65 mcc- folders.)

(3) Rewriting the code which processes platform folders; it is no
    longer used for user files, so the various code converting between
    Files and IntelliJ Virtual Files to compare paths is no longer
    needed.

This changeset also fixes the go-to-declaration element order shown
when jumping from XML resources. (This was recently fixed for
references from classes, but XML took a different code path which did
not include sorting.)

Change-Id: I31510b2d9705afe6d967a14fdbbd536702497c3f