Tweak layout editor selection for root nodes
Don't show visual bounds for the root node, unless it's in included context.
Also fix XML preview such that when you click to select a node, the
corresponding node is always selected. (This moves the caret to the beginning
character of the tag, but this is not considered "inside" the tag so in some
cases it leaves the previous node selected if they are adjacent.)

Change-Id: I659330e1f721c77a2098cccb7aec6b3b5b84a863