Project structure: use Gradle model for more complete UI
The Project Structure editors now query the Gradle model to get its
view of flavors and build types to show more complete information in
the editors. This lets users see what the default values will be
for properties that aren't explicitly set in the build files, and
also shows implicit default objects that are populated even if not
specified (such as debug and release build types). Additionally,
allows editing of the defaultConfig pseudo-flavor.

Bug: 68551
Bug: 61876
Bug: 61686

Change-Id: I4e370381d301d7ce2ddf5675fa05032f6e27ce61